+++ apps/dot-components-e2e/src/integration/List/List.spec.ts
@@ -32,9 +32,10 @@ describe('dot-components: List component with menu', () => {
     )
   );
 
-  it('should display the nested list as a menu', () => {
+  it('should display the nested list as a menu with icons shown', () => {
     cy.contains('Progressions').click();
     cy.get('.dot-menu').should('be.visible').contains('Package Progression');
+    cy.get('.dot-menu').get('.icon-block').should('be.visible');
   });
 });